Move these tests into their own file.
